pyproject.toml
[project]
name = "ipqs-tui"
version = "0.1.0"
description = "Textual TUI for the IPQualityScore API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "textual>=8.2.4",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -m 'not integration'"
asyncio_mode = "auto"
markers = [
    "integration: hits the real IPQS API (requires IPQS_API_KEY)",
]

[project.scripts]
ipqs-tui = "ipqs_tui.app:main"
